♻️ refactor:  マップとカードをひとまとめにした
#74
