--- lib/application/geolocator/state/current_map_spot_index.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-/// マップ表示の現在選択中のスポットのインデックス
-final currentMapSpotIndexProvider = StateProvider(
-  (_) => 0,
-  name: 'currentMapSpotIndexProvider',
-);
