+++ lib/presentation/page/stamp_rally/component/spot_map.dart
@@ -1,18 +1,124 @@
+import 'dart:async';
+
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:card_swiper/card_swiper.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart' show rootBundle;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
 
-import '../../../../application/geolocator/state/current_map_spot_index.dart';
+import '../../../../application/geolocator/state/current_poly_points.dart';
 import '../../../../application/geolocator/state/user_geolocator_position.dart';
+import '../../../../application/stamp_rally/state/pin_icon.dart';
+import '../../../../domain/entity/value_object/geo_location.dart';
 import '../../../../domain/repository/stamp_rally/entity/spot.dart';
+import '../../../../domain/repository/stamp_rally/entity/stamp_rally.dart';
+import '../../../../util/assets/assets.gen.dart';
 import '../../../component/async_value_handler.dart';
 
+/// マップ表示の現在選択中のスポットのインデックス
+final _currentMapSpotIndexProvider = StateProvider(
+  (_) => 0,
+);
+
+/// スポットのマップ
+class SpotMapView extends ConsumerStatefulWidget {
+  const SpotMapView({
+    super.key,
+    required this.stampRally,
+  });
+
+  final StampRally stampRally;
+
+  @override
+  _MapViewState createState() => _MapViewState();
+}
+
+class _MapViewState extends ConsumerState<SpotMapView> {
+  final Completer<GoogleMapController> mapController = Completer();
+
+  @override
+  Widget build(BuildContext context) {
+    // ピンアイコン画像
+    final icon = ref.watch(pinIconProvider);
+
+    // 初期表示時のカメラ位置
+    final spotIndex = ref.watch(_currentMapSpotIndexProvider);
+    final location = widget.stampRally.spots[spotIndex].location.toLatLng();
+
+    // 経路
+    final polylinePoints =
+        ref.watch(currentPolylinePointsProvider(widget.stampRally)).value ?? [];
+
+    // 選択中スポットインデックスを監視してマップ上のカメラを移動する
+    ref.listen(_currentMapSpotIndexProvider, (_, next) async {
+      final spot = widget.stampRally.spots[next];
+      final controller = await mapController.future;
+      await controller.showMarkerInfoWindow(MarkerId(spot.id));
+      await controller.animateCamera(
+        CameraUpdate.newLatLngZoom(
+          spot.location.toLatLng(),
+          17,
+        ),
+      );
+    });
+
+    return GoogleMap(
+      polylines: {
+        Polyline(
+          polylineId: const PolylineId('route'),
+          points: polylinePoints,
+          width: 4,
+          color: Theme.of(context).colorScheme.primaryContainer,
+        )
+      },
+      // hack ぬるぽで落ちる件のワークアラウンド
+      // https://github.com/flutter/flutter/issues/51064#issuecomment-588519913
+      minMaxZoomPreference: const MinMaxZoomPreference(0, 16),
+      myLocationEnabled: true,
+      zoomControlsEnabled: false,
+      onMapCreated: (controller) async {
+        mapController.complete(controller);
+        final value = await rootBundle.loadString(Assets.json.mapStyle);
+        final futureController = await mapController.future;
+        await futureController.setMapStyle(value); // Controllerを使ってMapをSetする
+      },
+      myLocationButtonEnabled: false,
+      // マップに表示するマーカーを生成する
+      markers: widget.stampRally.spots
+          .map(
+            (spot) => Marker(
+              infoWindow: InfoWindow(title: spot.title),
+              markerId: MarkerId(spot.id),
+              icon: icon,
+              position: LatLng(
+                spot.location.latitude,
+                spot.location.longitude,
+              ),
+            ),
+          )
+          .toSet(),
+      initialCameraPosition: CameraPosition(
+        target: location,
+        zoom: 14,
+      ),
+    );
+  }
+}
+
+extension _GeoLocationEx on GeoLocation {
+  /// GeoLocation => LatLng
+  LatLng toLatLng() => LatLng(
+        latitude,
+        longitude,
+      );
+}
+
 /// マップで使うスポットのカードを含むSwiper
-class MapSpotSwiper extends ConsumerWidget {
-  const MapSpotSwiper({
+class SpotMapSwiper extends ConsumerWidget {
+  const SpotMapSwiper({
     super.key,
     required this.spots,
   });
@@ -26,10 +132,10 @@ class MapSpotSwiper extends ConsumerWidget {
     return SizedBox(
       height: height,
       child: Swiper(
-        index: ref.watch(currentMapSpotIndexProvider),
+        index: ref.watch(_currentMapSpotIndexProvider),
         itemBuilder: (context, index) {
           final spot = spots[index];
-          return MapSpotCard(
+          return SpotMapCard(
             spot: spot,
           );
         },
@@ -37,7 +143,7 @@ class MapSpotSwiper extends ConsumerWidget {
         itemCount: spots.length,
         viewportFraction: 0.8,
         onIndexChanged: (index) {
-          ref.read(currentMapSpotIndexProvider.notifier).state = index;
+          ref.read(_currentMapSpotIndexProvider.notifier).state = index;
         },
       ),
     );
@@ -45,8 +151,8 @@ class MapSpotSwiper extends ConsumerWidget {
 }
 
 /// マップで使うスポットのカード
-class MapSpotCard extends ConsumerWidget {
-  const MapSpotCard({
+class SpotMapCard extends ConsumerWidget {
+  const SpotMapCard({
     super.key,
     required this.spot,
   });
